requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
boto3==1.26.66
botocore==1.29.66
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-stubs==1.14.0
django-stubs-ext==0.7.0
django-types==0.16.0
django-webpack-loader==0.7.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
joblib==1.2.0
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.1
openai==0.27.6
psycopg2==2.9.5
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
s3transfer==0.6.0
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.65.0
types-pytz==2022.7.1.0
types-PyYAML==6.0.12.3
typing-extensions==4.4.0
urllib3==1.26.13
yarl==1.9.2
